+++ Models/Api/User/UserFormModel.cs
@@ -6,8 +6,8 @@ namespace GymApp.Models.Api.User
         public string Username { get; set; }
         public string Password { get; set; }
         public int Weight { get; set; }
-        public float Height { get; set; }
+        public double Height { get; set; }
         public string Gender { get; set; }
-        public float Bmi { get; set; }
+        public double Bmi { get; set; }
     }
 }
