+++ Models/DataBase/UserMeasurement.cs
@@ -5,9 +5,9 @@ namespace GymApp.Models.DataBase
     public class UserMeasurement
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public int Type { get; set; }
-        public int Value { get; set; }
+        public long Value { get; set; }
         
 
         public User User { get; set; }
