+++ Services/TrainingService/ITrainingService.cs
@@ -8,8 +8,8 @@ namespace GymApp.Services.TrainingService
     {
         Training Create(TrainingFormModel model);
         
-        TrainingViewModel Fetch(int trainingId);
-        Training Update(TrainingFormModel model, int trainingId);
-        void Delete(int trainingId);
+        TrainingViewModel Fetch(long trainingId);
+        Training Update(TrainingFormModel model, long trainingId);
+        void Delete(long trainingId);
     }
 }
